+++ src/semgrep_agent/main.py
@@ -26,7 +26,6 @@ from semgrep_agent.exc import ActionFailure
 from semgrep_agent.meta import generate_meta_from_environment
 from semgrep_agent.meta import GitMeta
 from semgrep_agent.semgrep import SemgrepError
-from semgrep_agent.semgrep import SEMGREPIGNORE_ACTION
 from semgrep_agent.semgrep_app import Sapp
 from semgrep_agent.semgrep_app import SEMGREP_RULES_FILE
 from semgrep_agent.utils import get_aligned_command
@@ -357,7 +356,6 @@ def protected_main(
 
     committed_datetime = meta.commit.committed_datetime if meta.commit else None
     start_time = datetime.now()
-    semgrep.create_semgrepignore(sapp.scan.ignore_patterns)
     scan_context = semgrep.RunContext(
         config_specifier=config,
         committed_datetime=committed_datetime,
@@ -366,10 +364,9 @@ def protected_main(
         rewrite_rule_ids=(rewrite_rule_ids and not sapp.is_configured),
         enable_metrics=enable_metrics,
         timeout=(timeout if timeout > 0 else None),
-        action_ignores_path=str(SEMGREPIGNORE_ACTION),
+        requested_excludes=sapp.scan.ignore_patterns,
     )
     results = semgrep.scan(scan_context)
-    semgrep.SEMGREPIGNORE_ACTION.unlink()
     end_time = datetime.now()
 
     if not os.getenv("SEMGREP_AGENT_DEBUG"):
@@ -452,7 +449,7 @@ def protected_main(
             f"| {unit_len(results.findings.new_ignored, 'muted issue')} synced to the app",
             err=True,
         )
-        if len(results.findings.new_all) > 0:
+        if results.findings.new_all:
             click.echo(
                 f"| visit {publish_url}/orgs/{sapp.deployment_name}/findings?repo={meta.repo_name}",
                 err=True,