+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## Changed
 
+- .semgrepignore patterns are now handled by semgrep instead of semgrep-agent
+- baseline/diff scanning is now handled by semgrep instead of semgrep-agent
+
 ## Removed
 
 # 2022-02-24
@@ -17,6 +20,7 @@
 ## Added
 
 ## Changed
+
 - Instead of using env var SEMGREP_AGENT_OPT_IN_FEATURES to turn on autofix
   it is toggled using a setting in semgrep.dev (#517)